requirements.txt
# requirements.txt
uv
fastapi
uvicorn[standard]
streamlit
langchain
chromadb
sentence-transformers
pymupdf
pdfplumber
pytesseract
pdf2image
huggingface-hub
python-dotenv
langchain-nvidia-ai-endpoints
numpy
pillow
opencv-python
moviepy==1.0.3
speechrecognition
pydub
python-multipart
imageio
imageio-ffmpeg
openai
streamlit-markdown
requests
langchain-community
langgraph
